pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = "xXx"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
greenlet = "^3.0.3"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
